requirements.txt
selenium
selenium-findtext
pytest
pytest-dotenv